pyproject.toml
[tool.poetry]
name = "swole-v2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include = "tests"},
    {include = "cli"},
    {include = "swole_v2", from = "src"}
]

[tool.poetry.scripts]
db = "cli.db:db"
app = "cli.app:app"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
httpx = "^0.23.1"
python-multipart = "^0.0.5"
uvicorn = "^0.20.0"
sqlmodel = "^0.0.8"
pydantic = {extras = ["email"], version = "^1.10.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^0.21.0"
sqlalchemy-utils = "^0.39.0"
edgedb = "^1.2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
nox = "^2022.11.21"
flake8 = "^6.0.0"
black = "^22.12.0"
pytest-cov = "^4.0.0"
isort = "^5.11.2"
flake8-pyproject = "^1.2.2"
click = "^8.1.3"
pydantic-factories = "^1.17.0"
pytest-dotenv = "^0.5.2"
hypothesis = "^6.61.0"


[tool.poetry.group.stubs.dependencies]
types-passlib = "^1.7.7.3"
types-python-jose = "^3.3.4"
types-sqlalchemy = "^1.4.53.19"
types-click = "^7.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ============ EXRA TOOLS ============

[tool.mypy]
strict = true
disallow_untyped_decorators = false
disallow_subclassing_any = false
warn_return_any = false
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = [
    "nox.*",
    "fastapi.*",
    "pytest.*",
    "pydantic.*",
    "sqlmodel.*",
    "dotenv.*",
    "pydantic_factories.*",
    "uvicorn.*",
    "sqlalchemy_utils.*",
    "hypothesis.*",
    "faker.*",
    "edgedb.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
env_override_existing_values = 1
env_files = "test.env"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["swole_v2"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = ["if TYPE_CHECKING:"]

[tool.flake8]
max-line-length = 120
select = ["BLK","C","F","W","I"]
# E203,W503 don't play well with black (BLK)
ignore = ["E203","W503","E731","I100","I101","I202"]
exclude = "__init__.py"
application-import-names = ["swole_v2","tests"]
import-order-style = "google"

[tool.black]
line-length = 120

[tool.isort]
profile = "google"
order_by_type = true
force_single_line = false
multi_line_output = 3
include_trailing_comma = true
